training/requirements.txt
accelerate==0.15.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
blinker==1.4
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.0
cytoolz==0.11.0
efficientnet==1.1.1
et-xmlfile==1.1.0
filelock==3.9.0
fire==0.5.0
fonttools==4.25.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
h5py==3.7.0
huggingface-hub==0.11.1
importlib-metadata==4.12.0
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
kaggle==1.5.12
Keras-Applications==1.0.8
language-tool-python==2.7.1
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
mlxtend==0.21.0
munkres==1.1.4
numba==0.56.4
numpy==1.23.5
olefile==0.46
opencv-python==4.6.0.66
openpyxl==3.0.10
openslide-python==1.2.0
pandas==1.4.3
Pillow==9.2.0
pooch==1.6.0
protobuf==3.19.4
psutil==5.9.4
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyPDF2==2.11.1
python-slugify==6.1.2
PyYAML==6.0
q==2.7
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.1.2
seaborn==0.11.2
SimpleCRF==0.2.1.1
six==1.15.0
soundfile==0.11.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metal==0.5.0
termcolor==1.1.0
text-unidecode==1.3
tf-slim==1.1.0
threadpoolctl==3.1.0
timm==0.6.12
torch==1.13.0
torchaudio==0.13.0
torchfile==0.1.0
torchvision==0.14.0
tqdm==4.64.1
visdom==0.1.8.9
webencodings==0.5.1
Werkzeug==2.2.1
zipp==3.8.1
